Better visualization of decision trees (fixes #136) (#210)
* Added graphviz (dot) output for the rules() method in DecisionTree.

This change builds on the rules() method from a51d3f197602d624ffdcbc227c2569a0cd1d64f5.

The string generated can be send through the dot command to produce an image.

A test case for is added to ClassificationTreeTest.php

* removed spurious method

* Changes from review + composer check fixes.

The static analysis complained about Outcome lacking an "outcome" function, which was indeed the case. But all its implementing classes have such a function, so I took the liberty to add it to the interface.

* Changes from review.

Added image export (fixes #136)

Fixed bug in search function.